requirements.txt
# Core Flask Framework
Flask==3.1.1
Werkzeug==3.1.3

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
PyMySQL==1.1.1

# Authentication & Security
Flask-Login==0.6.3

# Email Support
Flask-Mail==0.10.0

# Environment Configuration
python-dotenv==1.1.1

# Data Processing & Export
pandas==2.3.1
xlsxwriter==3.2.5

# Additional Utilities
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
alembic==1.14.0
Mako==1.3.8
typing_extensions==4.12.2

# Development Dependencies (optional)
# pytest==8.3.4
# pytest-flask==1.3.0
# flask-testing==0.8.1